add react fronted

+++ app/Http/Controllers/API/AuthController.php
@@ -45,7 +45,7 @@ class AuthController extends Controller
         return response()->json([
             'access_token' => $token,
             'token_type'   => 'bearer',
-            'expires_in'   => (int) auth('api')->factory()->getTTL() * 60,
+            'expires_in'   => auth('api')->factory()->getTTL() * 60,
             'user'         => auth('api')->user()
         ]);
     }
